requirements.txt
py-clob-client==0.16.0
web3==6.14.0
requests==2.31.0
aiohttp==3.9.3
anthropic==0.25.0
sqlalchemy==2.0.28
python-dotenv==1.0.1
pyyaml==6.0.1
tenacity==8.2.3
structlog==24.1.0
sentence-transformers==2.7.0
numpy>=1.24.0
python-telegram-bot==21.0
PyGithub==2.1.1
